Move MaskValues to separate constants module
