+++ imod/common/utilities/constants.py
@@ -1,0 +1,16 @@
+import numpy as np
+
+
+from dataclasses import dataclass
+
+
+@dataclass
+class MaskValues:
+    """
+    Stores mask values for nodata. Special sentinel values can be stored in
+    here, such as the -9999.0 for MetaSWAP.
+    """
+
+    float = np.nan
+    integer = 0
+    msw_default = -9999.0
